+++ cache.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,22 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package cache provide a Cache interface and some implement engine
-// Usage:
-//
-// import(
-//   "github.com/beego/beego/v2/client/cache"
-// )
-//
-// bm, err := cache.NewCache("memory", `{"interval":60}`)
-//
-// Use it like this:
-//
-//	bm.Put("astaxie", 1, 10 * time.Second)
-//	bm.Get("astaxie")
-//	bm.IsExist("astaxie")
-//	bm.Delete("astaxie")
-//
 package cache
 
 import (
@@ -36,15 +20,6 @@ import (
 )
 
 // Cache interface contains all behaviors for cache adapter.
-// usage:
-//	cache.Register("file",cache.NewFileCache) // this operation is run in init method of file.go.
-//	c,err := cache.NewCache("file","{....}")
-//	c.Put("key",value, 3600 * time.Second)
-//	v := c.Get("key")
-//
-//	c.Incr("counter")  // now is 1
-//	c.Incr("counter")  // now is 2
-//	count := c.Get("counter").(int)
 type Cache interface {
 	// Get a cached value by key.
 	Get(ctx context.Context, key string) (interface{}, error)