+++ write_through.go
@@ -22,24 +22,24 @@ import (
 	berror "github.com/beego/beego-error/v2"
 )
 
-type WriteThoughCache struct {
+type WriteThroughCache struct {
 	Cache
 	storeFunc func(ctx context.Context, key string, val any) error
 }
 
-func NewWriteThoughCache(cache Cache, fn func(ctx context.Context, key string, val any) error) (*WriteThoughCache, error) {
+func NewWriteThroughCache(cache Cache, fn func(ctx context.Context, key string, val any) error) (*WriteThroughCache, error) {
 	if fn == nil || cache == nil {
 		return nil, berror.Error(InvalidInitParameters, "cache or storeFunc can not be nil")
 	}
 
-	w := &WriteThoughCache{
+	w := &WriteThroughCache{
 		Cache:     cache,
 		storeFunc: fn,
 	}
 	return w, nil
 }
 
-func (w *WriteThoughCache) Set(ctx context.Context, key string, val any, expiration time.Duration) error {
+func (w *WriteThroughCache) Set(ctx context.Context, key string, val any, expiration time.Duration) error {
 	err := w.storeFunc(ctx, key, val)
 	if err != nil {
 		return berror.Wrap(err, PersistCacheFailed, fmt.Sprintf("key: %s, val: %v", key, val))