+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - [Fix: refactoring the test and removing StartAndGC](https://github.com/beego/beego/pull/1)
 - [Fix: modify README](https://github.com/beego/beego/pull/2)
 - [Fix: Change the redis client library from redigo to go-redis](https://github.com/beego/beego-cache/pull/15)
+- [rename cache key to prefix of redis implementation](https://github.com/beego/beego-cache/pull/17)